+++ examples/stream/amqp/streamer.js
@@ -6,20 +6,11 @@ var messageCallback = function(content, field, properties) {
   console.log(content);
 };
 
-// Manual Config
-// var connectionParams = {
-//   client: 'your-client-id',
-//   secret: 'your-secret',
-//   host: 'host',
-//   port: 1883, // default for MQTT
-//   vhost: 'vhost'
-// };
-
 // Auto Config
 // You can also provide the endpointUrl to use a different end point, default is http://api.demo.spacebunny.io
 var connectionParams = {
   client: 'your-client-id',
-  secret: 'your-secret'
+  secret: 'your-secret',
 };
 
 // Auto Config with SSL
@@ -33,15 +24,28 @@ var connectionParams = {
 //   key: '/path/to/client_key.pem'
 // };
 
+// Manual Config
+// var connectionParams = {
+//   client: 'your-client-id',
+//   secret: 'your-secret',
+//   host: 'host',
+//   port: 61613, // default for MQTT
+//   vhost: 'vhost'
+// };
+
 // Stream hooks contains the stream name from which you want to collect data
 // and the callback which is invoked when receiving a message on that stream
 // the boolean cache option can be passed to specify the stream connection mode.
-// cache true (or missing) means that you want to read messages from the stream cache
-// cache false means that you want to read messages in a temporary queue that will be delete on disconnect
-// var streamHooks = [
-//   { stream: 'stream-name', callback: messageCallback },
-//   { stream: 'stream-name', cache: false, callback: messageCallback }
-// ];
+//
+// Options:
+// stream: represents the stream name
+// cache: (default true)
+//    true (or missing) means that you want to read messages from the stream cache
+//    false means that you want to read messages in a temporary queue that will be delete on disconnect
+var streamHooks = [
+  { stream: 'stream-name', callback: messageCallback },
+  { stream: 'stream-name-2', callback: messageCallback }
+];
 
 var streamClient = new AmqpStreamClient(connectionParams);
 streamClient.streamFrom(streamHooks).then(function(res) {