+++ CHANGELOG.md
@@ -2,12 +2,20 @@
 All notable changes to this project will be documented in this file.
 `Space Bunny Node SDK` adheres to [Semantic Versioning](http://semver.org/).
 
-- `0.1.x` Releases - [0.1.0](#210)
+- `0.1.x` Releases - [0.1.0]
+- `0.2.x` Releases - [0.2.0]
 
 ---
+
+## [0.2.0](https://github.com/space-bunny/node-sdk/releases/tag/v0.2.0)
+
+#### Added  
+
+- Ack / Nack handling for STOMP clients
+- support for DiscardMine and DiscardFromApi options
+
+## [0.1.0](https://github.com/space-bunny/node-sdk/releases/tag/v0.1.0)
 
 #### Added  
 - AMQP, STOMP, MQTT clients
 - AMQP, STOMP, MQTT stream clients
-
-## [0.1.0](https://github.com/space-bunny/releases/tag/0.1.0)